track_field_analyzer/requirements.txt
# Track & Field Form Analyzer - Dependencies
# Python 3.11+ required

# Core
streamlit>=1.28.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pandas>=2.0.0

# Config
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional)
mypy>=1.5.0
